+++ custom_components/gw_energypilot/frontend/gw-energy-pilot-v017.js
@@ -1,5 +1,5 @@
-import "./gw-energy-pilot-v016.js?v=1.2.0-beta.2-soc-end-sems2-beta-tests1";
-import "./gw-energy-pilot-settings-v016.js?v=1.2.0-beta.2-soc-end-sems2-beta-tests1";
+import "./gw-energy-pilot-v016.js?v=1.2.0-beta.3-load-forecast1";
+import "./gw-energy-pilot-settings-v016.js?v=1.2.0-beta.3-load-forecast1";
 
 const VERSION = "0.17";
 const PANEL_NAME = "gw-energypilot-panel";