+++ custom_components/gw_energypilot/frontend/gw-energy-pilot-v043.js
@@ -1,4 +1,4 @@
-import "./gw-energy-pilot-v042.js?v=1.2.0-beta.2-soc-end-sems2-beta-tests1";
+import "./gw-energy-pilot-v042.js?v=1.2.0-beta.3-load-forecast1";
 
 const VERSION = "0.43";
 const PANEL_NAME = "gw-energypilot-panel";